rag-agent/requirements.txt
fastapi
uvicorn
python-dotenv
openai
elevenlabs
llama-index-core
llama-index-readers-file
llama-index-embeddings-azure-openai
llama-index-vector-stores-faiss
faiss-cpu
python-multipart
jinja2
pydantic
llama-index-llms-azure-openai
numpy
tqdm
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pyttsx3
pydub
pandas
SpeechRecognition
gTTS
langdetect
ffmpeg-python
tiktoken
